refactor(directory): 페이지 디렉터리명 명시적으로 변경
