+++ src/app/lostfound/write/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const LostFoundWrite = () => {
+  return <LostFoundWritePage />;
+};
+
+export default LostFoundWrite;
