+++ src/page/forum-post/assets/CategoryArrow.tsx
