+++ src/page/forum-post/assets/CommentIcon.tsx
