+++ src/page/forum-post/assets/View.tsx
